updated ssx functions
