+++ packages/ssx-react/src/index.ts
@@ -1,1 +1,1 @@
-export * from './useSSX';
+export * from './ssx';
