--- packages/ssx-react/src/useSSX.ts
@@ -1,31 +1,0 @@
-import { SSX, SSXConfig } from "@spruceid/ssx";
-import { useEffect, useState } from 'react';
-import { useSigner } from 'wagmi';
-
-export function useSSX({ ssxConfig }: { ssxConfig?: SSXConfig }) {
-    const { data: signer, isSuccess: signerLoaded  } = useSigner()
-    const [ssx, setSSX] = useState<SSX>();
-
-    useEffect(() => {
-       async function initializeSSX() {
-            const { SSX } = await import('@spruceid/ssx');
-            const modifiedSSXConfig = {
-                ...ssxConfig,
-                providers: {
-                    ...ssxConfig?.providers,
-                    web3: { 
-                        driver: signer?.provider,
-                        ...ssxConfig?.providers?.web3,
-                    },
-                }
-            };
-            const ssxInstance = new SSX(modifiedSSXConfig);
-            setSSX(ssxInstance);
-        }
-        if (signerLoaded && signer) {
-            initializeSSX();
-        }
-    }, [signer, signerLoaded, ssxConfig]);   
-    
-    return { ssx }
-}
